usr/share/mops/requirements.txt
PyQt6>=6.4.0
slixmpp>=1.8.3
python-omemo>=1.1.0
slixmpp-omemo>=0.4.0
cryptography>=42.0.0
qasync>=0.25.0
